restructured cli relations
