+++ spel-katalog/cli/src/lib.rs
@@ -4,7 +4,6 @@ use ::clap::Parser;
 
 pub use self::{
     batch::Batch,
-    install_game::InstallGame,
     run::Run,
     subcmd::{SubCmdError, Subcmd, SubcmdCallbacks},
 };
@@ -12,7 +11,6 @@ pub use self::{
 mod batch;
 mod completions;
 mod init_config;
-mod install_game;
 mod run;
 mod skeleton;
 mod subcmd;
