+++ spel-katalog/src/main.rs
@@ -2,30 +2,13 @@ use ::std::io::{IsTerminal, Read};
 
 use ::mimalloc::MiMalloc;
 use ::spel_katalog::run as run_app;
-use ::spel_katalog_cli::{Cli, InstallGame, Subcmd, SubcmdCallbacks};
+use ::spel_katalog_cli::{Cli, Subcmd, SubcmdCallbacks};
 use ::spel_katalog_sink::SinkBuilder;
 
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
 
-fn init_log(target: Option<::env_logger::Target>) {
-    let mut log_builder = ::env_logger::builder();
-    log_builder.filter_level(::log::LevelFilter::Info);
-
-    if let Some(target) = target {
-        log_builder.target(target).init();
-    } else {
-        log_builder.init();
-    }
-}
-
-fn other() -> ::color_eyre::Result<()> {
-    init_log(None);
-    Ok(())
-}
-
 fn run(cli: ::spel_katalog_cli::Run) -> ::color_eyre::Result<()> {
-    init_log(None);
     let keep_terminal = cli.keep_terminal;
     run_app(cli, SinkBuilder::Inherit, None)?;
 
@@ -38,18 +21,17 @@ fn run(cli: ::spel_katalog_cli::Run) -> ::color_eyre::Result<()> {
     Ok(())
 }
 
-fn install_game(install_game: InstallGame) -> ::color_eyre::Result<()> {
-    init_log(None);
-    ::spel_katalog_install::install_game(install_game)
-}
-
 fn main() -> ::color_eyre::Result<()> {
     ::color_eyre::install()?;
     let cli = Cli::parse();
     let cmd = Subcmd::from(cli);
-    cmd.perform(SubcmdCallbacks {
-        run,
-        other,
-        install_game,
-    })
+    let mut log_builder = ::env_logger::builder();
+    log_builder.filter_level(::log::LevelFilter::Info);
+
+    if let Some(target) = None {
+        log_builder.target(target).init();
+    } else {
+        log_builder.init();
+    }
+    cmd.perform(SubcmdCallbacks { run })
 }
